Makes ThreadLocalServerAndClientSpanState public and polishes use
ServerAndClientSpanState is awkward to use without examples. This shows
use of the initial state value `ServerSpan.EMPTY`, and polishes Brave's
builder, which accidentally implied creating state twice.

This change is paid for by removing `SimpleServerAndClientSpanState`,
which was unusable dead code due to package protection.

This also makes ThreadLocalServerAndClientSpanState public so it can be
used directly in Spring.
